pyproject.toml
[tool.poetry]
name = "university timetabling automation of schedule"
version = "0.1"
description = ""
authors = ["Sagat Akimbay <maroccocombo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
python-dateutil = "^2.9.0.post0"
structlog = "^25.4.0"
django-prometheus = "^2.4.0"
marshmallow = "^4.0.0"
xlsxwriter = "^3.2.5"
reportlab = "^4.4.2"
openpyxl = "^3.1.5"
django-filter = "^25.1"
django-cors-headers = "^4.7.0"
python-dotenv = "^1.1.0"
pandas = "^2.3.0"
django-import-export = "^4.3.7"
djangorestframework = "^3.16.0"
optapy = "^9.37.0b0"
celery = {extras = ["redis"], version = "^5.5.3"}
psycopg2-binary = "^2.9.10"
redis = "5.0.4"
django-redis = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
